Move transaction checking to web3 poller
